Reorganize modules: flatten pretty_printing, rename pane_printing
