--- src/pane_printing/mod.rs
@@ -1,12 +1,0 @@
-mod divvy;
-mod pane_notation;
-mod pane_print;
-mod plain_text;
-mod pretty_window;
-mod render_options;
-
-pub use pane_notation::{DocLabel, PaneNotation, PaneSize};
-pub use pane_print::{pane_print, PaneError};
-pub use plain_text::PlainText;
-pub use pretty_window::PrettyWindow;
-pub use render_options::{FocusSide, RenderOptions, WidthStrategy};
